Allows for @ in urls for profile edits
Use Case: When users want to edit their profile and add a link with an @ sign in it, it will say 'The field name url format is invalid'. If someone wanted their website url to be a Whaleshares or Steem link, it won't let them. This PR adds support for @ signs in urls.